requirements.txt
requests
ollama
python-dotenv
rich
